move produce/fetch tests into separate file
